requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.3
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.9.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Markdown==3.4.4
phonenumbers==8.13.20
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
typing-extensions==4.8.0
uritemplate==4.1.1
zipp==3.17.0
